Started Setup instructions
I started working on the instructions for the setup. I made two comments in there, that shouldn't stay long term:
First I wasn't sure if there is a specific version of Python people might need. If so, we should probably add that.
Secondly I remarked that we do not have the mentioned requirements.txt file yet, I'm currently not entirely sure where to get the overview over our used packages